+++ src/main/java/Dijkstra.java
@@ -9,7 +9,7 @@ import main.java.commom.graph.Node;
 
 public class Dijkstra {
    
-   private HashMap<Node, Double> dists;
+   private HashMap<Node, Long> dists;
    private HashMap<Node, Boolean> isNodeComplete;
    
    private void dijkstra(Node pivot) {
@@ -19,7 +19,7 @@ public class Dijkstra {
       while(!queue.isEmpty()){
          NodeDist currentNodeDist = queue.remove();
          Node currentNode = currentNodeDist.node;
-         double currentDist = currentNodeDist.dist;
+         long currentDist = currentNodeDist.dist;
 
          isNodeComplete.put(currentNode, true);
 
@@ -27,7 +27,7 @@ public class Dijkstra {
             if(isNodeComplete.get(edge.nodeTo)) continue;
             
             Node nodeTo = edge.nodeTo;
-            double newDist = currentDist + edge.weight;
+            long newDist = currentDist + edge.weight;
             if(newDist < dists.get(nodeTo)){
                dists.put(nodeTo, newDist);
